Simple Date Counter
This is an extremely simple date counter application. Its only function is to display what the date will be after a given number of dates. It is simply a learning challenge for me to consolidate what i have recently been learning about controlled components. Which is why it is all written in a single App file as opposed to being in individual component files. Up to now I have not published these  "Throw away projects" I create as practice to git before, however it was pointed out to me that i should going forward.